requirements.txt
PyQt5
mysql-connector-python
pandas
numpy
scikit-learn
statsmodels
openpyxl